Move poll module to bau package
