+++ korben/korben/bau/__init__.py
@@ -1,4 +1,6 @@
-from . import webserver
+from . import webserver, poll
 
 def main():
     webserver.start()
+
+COMMANDS = {'poll': poll.main}
